Working on the clustering logic.
git-svn-id: http://ikube.googlecode.com/svn/trunk@1255 89a0fb20-889e-aefa-90f6-7c296b20d5cf

+++ code/core/src/main/java/ikube/IConstants.java
@@ -179,7 +179,7 @@ public interface IConstants {
 	String ROLE_USER = "ROLE_USER";
 	String ROLE_ADMIN = "ROLE_ADMIN";
 	
-	long MILLION = 1000000;
+	int MILLION = 1000000;
 	
 	String DELIMITER_CHARACTERS = ";,|:";
 	String LINE_NUMBER = "lineNumber";
